requirements.txt

pylacrosse==0.4
paho-mqtt==1.4.0
wheel==0.29.0
sdnotify==0.3.1
colorama==0.3.9
Unidecode==0.4.21
